tui/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blonde-cli"
version = "0.1.3"
description = "AI-powered code assistant CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name="Amardeep", email="amarbavi12345@gmail.com" }
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-magic>=0.4.27",
    "llama-cpp-python>=0.2.0",
    "huggingface-hub>=0.19.0",
    "GitPython>=3.1.40",
    "chromadb>=0.4.0",
    "keyring>=24.0.0",
    "mcp"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]
tree-sitter = [
    "tree-sitter>=0.20.0",
    "tree-sitter-languages>=1.8.0"
]
gpu = ["llama-cpp-python[cublas]"]
metal = ["llama-cpp-python[metal]"]

[project.urls]
Homepage = "https://github.com/YOUR_GITHUB/blonde-cli"
Repository = "https://github.com/YOUR_GITHUB/blonde-cli"

[project.scripts]
blnd = "cli:app"
